Added Collinear Points
Feature extraction is a sub-problem of pattern recognition and is also used in areas such as statistical analysis, computer vision, and image processing. For example, an image processing problem may use a feature extraction algorithm to identify particular shapes or regions in a digitized image. In this program, a set of points in two-dimensional space, are to be identified if every subset of four or more points are collinear.